+++ secretspec/src/provider/keyring.rs
@@ -183,7 +183,7 @@ impl Provider for KeyringProvider {
         &["field"]
     }
 
-    fn entry_coordinates<'a>(
+    fn configured_entry_coordinates<'a>(
         &self,
         addr: Address<'a>,
     ) -> Result<std::borrow::Cow<'a, crate::config::NativeAddress>> {